Update class.loader.php
Updates for checkDirectory and checkFile function to have a second boolean parameter isRootPath. By default isRootPath is set to false. This is very useful if you have a direct physical path that is not within the physical path of the mvc. Just pass the value of true with the actual physical path and your good to go!